Faster unit tests
