+++ src/test/datascience/helpers.ts
@@ -1,0 +1,35 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+'use strict';
+
+import { IDataScienceSettings } from '../../client/common/types';
+
+// The default base set of data science settings to use
+export function defaultDataScienceSettings(): IDataScienceSettings {
+    return {
+        allowImportFromNotebook: true,
+        jupyterLaunchTimeout: 10,
+        jupyterLaunchRetries: 3,
+        enabled: true,
+        jupyterServerURI: 'local',
+        // tslint:disable-next-line: no-invalid-template-strings
+        notebookFileRoot: '${fileDirname}',
+        changeDirOnImportExport: false,
+        useDefaultConfigForJupyter: true,
+        jupyterInterruptTimeout: 10000,
+        searchForJupyter: true,
+        showCellInputCode: true,
+        collapseCellInputCodeByDefault: true,
+        allowInput: true,
+        maxOutputSize: 400,
+        errorBackgroundColor: '#FFFFFF',
+        sendSelectionToInteractiveWindow: false,
+        variableExplorerExclude: 'module;function;builtin_function_or_method',
+        codeRegularExpression: '^(#\\s*%%|#\\s*\\<codecell\\>|#\\s*In\\[\\d*?\\]|#\\s*In\\[ \\])',
+        markdownRegularExpression: '^(#\\s*%%\\s*\\[markdown\\]|#\\s*\\<markdowncell\\>)',
+        enablePlotViewer: true,
+        runStartupCommands: '',
+        debugJustMyCode: true
+    };
+}
